action_projects/media-integrity/requirements.txt
# Media Integrity — Core Dependencies
# Install: pip install -r requirements.txt

# Image forensics
Pillow>=9.0.0
exifread>=2.3.2

# Text analysis
scikit-learn>=1.0.0  # optional but recommended
numpy>=1.20.0

# Video deepfake detection
opencv-python>=4.5.0
# dlib>=19.22.0  # optional (requires CMake, boost) — install manually if needed

# Network analysis
networkx>=2.6.0

# REST API
Flask>=2.0.0

# Optional but useful
requests>=2.25.0  # for external API calls (WHOIS, etc.)
python-dateutil>=2.8.0

# Testing
pytest>=6.2.0

# Note: Some dependencies are optional (dlib). Core functionality works without them.
# If optional module import fails, Media Integrity falls back to simpler heuristics.
